requirements.txt
scikit-video
pathlib
py-yaml
